requirements.txt
torch>=2.3.0
torchvision>=0.18.0
transformers==4.46.3
datasets>=2.21.0
accelerate>=0.34.0
decord>=0.6.0
opencv-python>=4.10.0
pillow>=10.4.0
pyyaml>=6.0.2
tqdm>=4.66.0
sentencepiece>=0.2.0
protobuf>=5.27.0
safetensors>=0.4.3
numpy>=1.26.0
pandas>=2.2.0
huggingface_hub>=0.24.0
av>=12.0.0
